Connected frontend with backend validation API
